back/requirements.txt
typesense==0.15.1
# NumPy with version pinned to 1.x for compatibility
numpy==1.24.3
# PyTorch and transformers dependencies with specific versions
torch==2.0.1
# Pin huggingface_hub to a version that has cached_download
huggingface_hub==0.16.4
transformers==4.30.2
sentence-transformers==2.2.2
# Other dependencies
markdown==3.5
python-dotenv==1.0.1
tqdm==4.66.2
click==8.1.7
fastapi==0.110.0
uvicorn==0.27.0
python-socketio==5.10.0
websockets==12.0
requests==2.31.0
psutil==5.9.5
